An animation with 2 cameras
i made a if-ish animation with the camera. The camera will go behind the player.

The only thing a need to do is to play the animation only when pressed play. The Intro camera will disable en the main camera will be enabled.

ps. I have a bug with when the into camera is done playing..